demo-app/requirements.txt
flask
requests
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-sdk
opentelemetry-api
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests